requirements.txt
streamlit
praw
pandas
nltk
openai
python-dotenv
numpy
altair
